chore(research): excluding unused fields from /api/articles endpoint
By not returning the article's `content`, we're able to improve the UI
loading time significantly since the payload being trasnported is
smaller.

Also removed everything that's not being used by the UI when it fetchs
the list of all articles. If needed, in the future we can reintroduce
the ones required